+++ app/app.go
@@ -69,29 +69,36 @@ import (
 	authtx "github.com/cosmos/cosmos-sdk/x/auth/tx"
 	"github.com/cosmos/cosmos-sdk/x/crisis"
 
-	"github.com/osmosis-labs/osmosis/v20/app/keepers"
-	"github.com/osmosis-labs/osmosis/v20/app/upgrades"
-	v10 "github.com/osmosis-labs/osmosis/v20/app/upgrades/v10"
-	v11 "github.com/osmosis-labs/osmosis/v20/app/upgrades/v11"
-	v12 "github.com/osmosis-labs/osmosis/v20/app/upgrades/v12"
-	v13 "github.com/osmosis-labs/osmosis/v20/app/upgrades/v13"
-	v14 "github.com/osmosis-labs/osmosis/v20/app/upgrades/v14"
-	v15 "github.com/osmosis-labs/osmosis/v20/app/upgrades/v15"
-	v16 "github.com/osmosis-labs/osmosis/v20/app/upgrades/v16"
-	v17 "github.com/osmosis-labs/osmosis/v20/app/upgrades/v17"
-	v18 "github.com/osmosis-labs/osmosis/v20/app/upgrades/v18"
-	v19 "github.com/osmosis-labs/osmosis/v20/app/upgrades/v19"
-	v20 "github.com/osmosis-labs/osmosis/v20/app/upgrades/v20"
-	v21 "github.com/osmosis-labs/osmosis/v20/app/upgrades/v21"
-	v3 "github.com/osmosis-labs/osmosis/v20/app/upgrades/v3"
-	v4 "github.com/osmosis-labs/osmosis/v20/app/upgrades/v4"
-	v5 "github.com/osmosis-labs/osmosis/v20/app/upgrades/v5"
-	v6 "github.com/osmosis-labs/osmosis/v20/app/upgrades/v6"
-	v7 "github.com/osmosis-labs/osmosis/v20/app/upgrades/v7"
-	v8 "github.com/osmosis-labs/osmosis/v20/app/upgrades/v8"
-	v9 "github.com/osmosis-labs/osmosis/v20/app/upgrades/v9"
-	_ "github.com/osmosis-labs/osmosis/v20/client/docs/statik"
-	"github.com/osmosis-labs/osmosis/v20/x/mint"
+	protorevtypes "github.com/osmosis-labs/osmosis/v21/x/protorev/types"
+
+	"github.com/osmosis-labs/osmosis/v21/app/keepers"
+	"github.com/osmosis-labs/osmosis/v21/app/upgrades"
+	v10 "github.com/osmosis-labs/osmosis/v21/app/upgrades/v10"
+	v11 "github.com/osmosis-labs/osmosis/v21/app/upgrades/v11"
+	v12 "github.com/osmosis-labs/osmosis/v21/app/upgrades/v12"
+	v13 "github.com/osmosis-labs/osmosis/v21/app/upgrades/v13"
+	v14 "github.com/osmosis-labs/osmosis/v21/app/upgrades/v14"
+	v15 "github.com/osmosis-labs/osmosis/v21/app/upgrades/v15"
+	v16 "github.com/osmosis-labs/osmosis/v21/app/upgrades/v16"
+	v17 "github.com/osmosis-labs/osmosis/v21/app/upgrades/v17"
+	v18 "github.com/osmosis-labs/osmosis/v21/app/upgrades/v18"
+	v19 "github.com/osmosis-labs/osmosis/v21/app/upgrades/v19"
+	v20 "github.com/osmosis-labs/osmosis/v21/app/upgrades/v20"
+	v21 "github.com/osmosis-labs/osmosis/v21/app/upgrades/v21"
+	v3 "github.com/osmosis-labs/osmosis/v21/app/upgrades/v3"
+	v4 "github.com/osmosis-labs/osmosis/v21/app/upgrades/v4"
+	v5 "github.com/osmosis-labs/osmosis/v21/app/upgrades/v5"
+	v6 "github.com/osmosis-labs/osmosis/v21/app/upgrades/v6"
+	v7 "github.com/osmosis-labs/osmosis/v21/app/upgrades/v7"
+	v8 "github.com/osmosis-labs/osmosis/v21/app/upgrades/v8"
+	v9 "github.com/osmosis-labs/osmosis/v21/app/upgrades/v9"
+	_ "github.com/osmosis-labs/osmosis/v21/client/docs/statik"
+	"github.com/osmosis-labs/osmosis/v21/ingest"
+	"github.com/osmosis-labs/osmosis/v21/x/mint"
+
+	"github.com/osmosis-labs/osmosis/v21/ingest/sqs"
+
+	"github.com/osmosis-labs/osmosis/v21/ingest/sqs/pools/common"
 )
 
 const appName = "OsmosisApp"
@@ -109,7 +116,7 @@ var (
 	maccPerms = moduleAccountPermissions
 
 	// module accounts that are allowed to receive tokens.
-	allowedReceivingModAcc = map[string]bool{}
+	allowedReceivingModAcc = map[string]bool{protorevtypes.ModuleName: true}
 
 	// TODO: Refactor wasm items into a wasm.go file
 	// WasmProposalsEnabled enables all x/wasm proposals when it's value is "true"
@@ -244,6 +251,31 @@ func NewOsmosisApp(
 		wasmOpts,
 		app.BlockedAddrs(),
 	)
+
+	// Initialize the ingest manager for propagating data to external sinks.
+	app.IngestManager = ingest.NewIngestManager()
+
+	sqsConfig := sqs.NewConfigFromOptions(appOpts)
+
+	// Initialize the SQS ingester if it is enabled.
+	if sqsConfig.IsEnabled {
+		sqsKeepers := common.SQSIngestKeepers{
+			GammKeeper:         app.GAMMKeeper,
+			CosmWasmPoolKeeper: app.CosmwasmPoolKeeper,
+			BankKeeper:         app.BankKeeper,
+			ProtorevKeeper:     app.ProtoRevKeeper,
+			PoolManagerKeeper:  app.PoolManagerKeeper,
+			ConcentratedKeeper: app.ConcentratedLiquidityKeeper,
+		}
+
+		sqsIngester, err := sqsConfig.Initialize(appCodec, sqsKeepers)
+		if err != nil {
+			panic(err)
+		}
+
+		// Set the sidecar query server ingester to the ingest manager.
+		app.IngestManager.RegisterIngester(sqsIngester)
+	}
 
 	// TODO: There is a bug here, where we register the govRouter routes in InitNormalKeepers and then
 	// call setupHooks afterwards. Therefore, if a gov proposal needs to call a method and that method calls a
@@ -392,6 +424,8 @@ func (app *OsmosisApp) BeginBlocker(ctx sdk.Context, req abci.RequestBeginBlock)
 
 // EndBlocker application updates every end block.
 func (app *OsmosisApp) EndBlocker(ctx sdk.Context, req abci.RequestEndBlock) abci.ResponseEndBlock {
+	// Process the block and ingest data into various sinks.
+	app.IngestManager.ProcessBlock(ctx)
 	return app.mm.EndBlock(ctx, req)
 }