+++ x/superfluid/keeper/params.go
@@ -1,7 +1,7 @@
 package keeper
 
 import (
-	"github.com/osmosis-labs/osmosis/v20/x/superfluid/types"
+	"github.com/osmosis-labs/osmosis/v21/x/superfluid/types"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 )